move iohanalyzer to logging, add save in node env
